+++ tablesoft-net/TableSoft/TableSoft/frmGestionarEquipo.cs
@@ -82,12 +82,31 @@ namespace TableSoft
             }
             equipo.nombre = txtNombre.Text;
             equipo.descripcion = txtDescripcion.Text;
-            if (equipoDAO.insertarEquipo(equipo) > 0)
+            if (MessageBox.Show("¿Desea crear el registro?", "Crear Equipo", MessageBoxButtons.YesNo) == DialogResult.Yes)
+            {
+                if (equipoDAO.insertarEquipo(equipo) > 0)
+                {
+                    MessageBox.Show(
+                    "Se ha creado el registro exitosamente",
+                    "Registro exitoso",
+                    MessageBoxButtons.OK, MessageBoxIcon.Information
+                    );
+                }
+                else
+                {
+                    MessageBox.Show(
+                    "No se ha creado el registro",
+                    "Registro no realizado",
+                    MessageBoxButtons.OK, MessageBoxIcon.Information
+                    );
+                }
+            }
+            else
             {
                 MessageBox.Show(
-                    "Se ha guardado el registro.",
-                    "Guardado exitoso",
-                    MessageBoxButtons.OK, MessageBoxIcon.Information
+                "No se ha creado el registro",
+                "Registro no realizado",
+                MessageBoxButtons.OK, MessageBoxIcon.Information
                 );
             }
             txtIDEquipo.Text = equipo.equipoId.ToString();
@@ -96,12 +115,31 @@ namespace TableSoft
 
         private void btnEliminar_Click(object sender, EventArgs e)
         {
-            if (equipoDAO.eliminarEquipo(equipo) > -1)
+            if (MessageBox.Show("¿Desea eliminar el registro?", "Eliminar Equipo", MessageBoxButtons.YesNo) == DialogResult.Yes)
             {
-                MessageBox.Show(
-                    "Se ha eliminado el registro.",
+                if (equipoDAO.eliminarEquipo(equipo) > -1)
+                {
+                    MessageBox.Show(
+                    "Se ha eliminado el registro exitosamente",
                     "Eliminación exitosa",
                     MessageBoxButtons.OK, MessageBoxIcon.Information
+                    );
+                }
+                else
+                {
+                    MessageBox.Show(
+                    "No se eliminó el registro",
+                    "Eliminación no realizada",
+                    MessageBoxButtons.OK, MessageBoxIcon.Information
+                    );
+                }
+            }
+            else
+            {
+                MessageBox.Show(
+                "No se eliminó el registro",
+                "Eliminación no realizada",
+                MessageBoxButtons.OK, MessageBoxIcon.Information
                 );
             }
             this.DialogResult = DialogResult.OK;
@@ -152,12 +190,31 @@ namespace TableSoft
             }
             equipo.nombre = txtNombre.Text;
             equipo.descripcion = txtDescripcion.Text;
-            if (equipoDAO.actualizarEquipo(equipo) > -1)
+            if (MessageBox.Show("¿Desea actualizar el registro?", "Actualizar Equipo", MessageBoxButtons.YesNo) == DialogResult.Yes)
             {
-                MessageBox.Show(
-                    "Se ha actualizado el registro.",
+                if (equipoDAO.actualizarEquipo(equipo) > -1)
+                {
+                    MessageBox.Show(
+                    "Se ha actualizado el registro exitosamente",
                     "Actualización exitosa",
                     MessageBoxButtons.OK, MessageBoxIcon.Information
+                    );
+                }
+                else
+                {
+                    MessageBox.Show(
+                    "No se ha realizado la actualización",
+                    "Actualización no realizada",
+                    MessageBoxButtons.OK, MessageBoxIcon.Information
+                    );
+                }
+            }
+            else
+            {
+                MessageBox.Show(
+                "No se ha realizado la actualización",
+                "Actualización no realizada",
+                MessageBoxButtons.OK, MessageBoxIcon.Information
                 );
             }
             this.DialogResult = DialogResult.OK;
